refactor: Extract JudgeConfig interface and a corresponding implementation into the impl package
